Added Next Explanation script, changed Next scenario script
Next explanation brings from scenario to explanation

Next scenario brings from explanation to scenario
