+++ Assets/Scenes/Explanation12.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3e45367dbb49239479c055611a043f56
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
