+++ Assets/Scripts/NextExplanation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 14e55b5b30e835142af44fe67d640c8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
