+++ Assets/Scripts/NextScenario.cs
@@ -5,69 +5,8 @@ using UnityEngine.SceneManagement;
 
 public class NextScenario : MonoBehaviour
 {
-
-   
-    public void choice1()
+    public void next()
     {
-        //store choice 
-        Player.choiceArr[Player.scenarios -1] = new Choice(Player.scenarios, 1); 
-
-        //load new scene 
-        Player.ageUp();
-        if (Player.age != "Dead")
-        {
-            string scene = "Scenario" + Player.scenarios;
-            SceneManager.LoadScene(scene);
-        }
-        else
-        {
-            SceneManager.LoadScene("End");
-        }
-
-     }
-
-
-    public void choice2()
-    {
-        //store choice 
-        Player.choiceArr[Player.scenarios - 1] = new Choice(Player.scenarios, 2);
-
-        //load new scene 
-        Player.ageUp();
-        if (Player.age != "Dead")
-        {
-            string scene = "Scenario" + Player.scenarios;
-            SceneManager.LoadScene(scene);
-        } else
-        {
-            SceneManager.LoadScene("End");
-        }
-    }
-
-    public void choice3()
-    {
-        //store choice 
-        Player.choiceArr[Player.scenarios - 1] = new Choice(Player.scenarios, 3);
-
-        //load new scene 
-        Player.ageUp();
-        if (Player.age != "Dead")
-        {
-            string scene = "Scenario" + Player.scenarios;
-            SceneManager.LoadScene(scene);
-        }
-        else
-        {
-            SceneManager.LoadScene("End");
-        }
-    }
-
-    public void choice4()
-    {
-        //store choice 
-        Player.choiceArr[Player.scenarios - 1] = new Choice(Player.scenarios, 4);
-
-        //load new scene 
         Player.ageUp();
         if (Player.age != "Dead")
         {
